clippy and renames
