+++ crates/bevy_animation_graph_core/src/utils/mod.rs
@@ -1,6 +1,7 @@
 pub mod delaunay;
 pub mod geometry;
 pub mod loading;
+pub mod sorted_map;
 
 use std::path::PathBuf;
