+++ crates/bevy_animation_graph_editor/src/ui/generic_widgets/mod.rs
@@ -6,7 +6,6 @@ pub mod body_mode;
 pub mod bone_id;
 pub mod bone_id_via_path;
 pub mod bone_mask;
-pub mod data_only_spec;
 pub mod data_spec_widget;
 pub mod data_value;
 pub mod entity_path;
@@ -24,6 +23,7 @@ pub mod popup;
 pub mod popup_asset_picker;
 pub mod quat;
 pub mod ragdoll_config;
+pub mod sorted_map;
 pub mod string_picker;
 pub mod tree;
 pub mod u32_flags;
