+++ crates/bevy_animation_graph_editor/src/ui/reflect_widgets/mod.rs
@@ -3,25 +3,22 @@ use bevy_animation_graph::core::{animation_graph::PinId, edge_data::DataSpecWith
 use crate::ui::{
     reflect_lib::WidgetRegistry,
     reflect_widgets::{
-        data_only_spec::DataOnlySpecReflectWidget, data_value::DataValueReflectWidget,
-        hashmap::HashMapReflectWidget, ragdoll_config::RagdollConfigReflectWidget,
+        data_value::DataValueReflectWidget, hashmap::HashMapReflectWidget,
+        ragdoll_config::RagdollConfigReflectWidget, sorted_map::SortedMapReflectWidget,
         vec3::Vec3ReflectWidget,
     },
 };
 
-pub mod data_only_spec;
 pub mod data_value;
 pub mod hashmap;
 pub mod ragdoll_config;
+pub mod sorted_map;
 pub mod vec3;
 
 pub fn register_reflect_widgets(registry: &mut WidgetRegistry) {
     registry
         .add(HashMapReflectWidget::<String, String>::default())
-        .add(DataOnlySpecReflectWidget::<
-            PinId,
-            DataSpecWithOptionalDefault,
-        >::default())
+        .add(SortedMapReflectWidget::<PinId, DataSpecWithOptionalDefault>::default())
         .add(DataValueReflectWidget::default())
         .add(Vec3ReflectWidget::default())
         .add(RagdollConfigReflectWidget::default());
