+++ crates/bevy_animation_graph_editor/src/ui/reflect_widgets/sorted_map.rs
@@ -1,25 +1,25 @@
 use std::marker::PhantomData;
 
 use bevy::reflect::Reflect;
-use bevy_animation_graph::core::context::spec_context::DataOnlySpec;
+use bevy_animation_graph::core::utils::sorted_map::SortedMap;
 
 use crate::ui::{
-    generic_widgets::data_only_spec::DataOnlySpecWidget,
+    generic_widgets::sorted_map::SortedMapWidget,
     reflect_lib::{ReflectWidget, ReflectWidgetContext},
 };
 
 #[derive(Default)]
-pub struct DataOnlySpecReflectWidget<K, V> {
+pub struct SortedMapReflectWidget<K, V> {
     __k: PhantomData<K>,
     __v: PhantomData<V>,
 }
 
-impl<K, V> ReflectWidget for DataOnlySpecReflectWidget<K, V>
+impl<K, V> ReflectWidget for SortedMapReflectWidget<K, V>
 where
     K: Reflect + Default + Clone + std::hash::Hash + std::fmt::Debug + Ord + Send + Sync + 'static,
     V: Reflect + Default + PartialEq + Clone + Send + Sync + 'static,
 {
-    type Target = DataOnlySpec<K, V>;
+    type Target = SortedMap<K, V>;
 
     fn draw(
         &self,
@@ -27,7 +27,7 @@ where
         value: &mut Self::Target,
         ctx: &ReflectWidgetContext,
     ) -> egui::Response {
-        DataOnlySpecWidget::new(value)
+        SortedMapWidget::new(value)
             .salted("reflect data only spec widget")
             .show(ui, |ui, k| ctx.draw(ui, k), |ui, v| ctx.draw(ui, v))
     }
